Add files via upload
This script is only viable when you have IMX252 and stream the video over gstreamer on Linux Host PC.

This takes a snapshot every frame possible and computes the brightness (level) by taking a mean of every image bit. The image is in greyscale, so each bit has a value of 0 to 255, 0 being black and 255 being white. Then, it is rescaled to 0 to 1.

As the level is too low, the script will increase the exposure and gain if needed. If it is too bright, it goes back to the default setting which is exposure=160, gain=0. It never goes beyond exposure>30000 and gain>400.